pyproject.toml
[project]
name = "mr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "great-tables>=0.17.0",
    "jupyter>=1.1.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
